Refactor wizards architecture for Slim and Classic separation
Separated Slim and Classic wizard implementations to improve clarity and modularity. Removed obsolete interfaces and cleaned up namespace organization for better maintainability. Updated project files and templates accordingly.
