+++ samples/TestApp/TestApp.Browser/Program.cs
@@ -3,16 +3,17 @@ using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Browser;
 using Avalonia.ReactiveUI;
-using TestApp;
 
 [assembly: SupportedOSPlatform("browser")]
+
+namespace TestApp.Browser;
 
 internal partial class Program
 {
     private static async Task Main(string[] args) => await BuildAvaloniaApp()
-            .WithInterFont()
-            .UseReactiveUI()
-            .StartBrowserAppAsync("out");
+        .WithInterFont()
+        .UseReactiveUI()
+        .StartBrowserAppAsync("out");
 
     public static AppBuilder BuildAvaloniaApp()
         => AppBuilder.Configure<App>();
