+++ samples/TestApp/TestApp/Samples/ControlsNew/Wizard/Pages/FirstPageViewModel.cs
@@ -4,19 +4,19 @@ using CSharpFunctionalExtensions;
 using ReactiveUI.SourceGenerators;
 using ReactiveUI.Validation.Extensions;
 using ReactiveUI.Validation.Helpers;
-using Zafiro.Avalonia.Controls.Wizards.Builder;
+using Zafiro.UI.Wizards.Classic.Builder;
 
 namespace TestApp.Samples.ControlsNew.Wizard.Pages;
 
 public partial class FirstPageViewModel : ReactiveValidationObject, IStep
 {
+    [Reactive] private int? number;
+
     public FirstPageViewModel()
     {
-        this.ValidationRule(x => x.Number, i => i  % 2 == 0, "Number must be even" );
+        this.ValidationRule(x => x.Number, i => i % 2 == 0, "Number must be even");
     }
 
-    [Reactive] private int? number;
-    
     public IObservable<bool> IsValid => this.IsValid();
     public IObservable<bool> IsBusy => Observable.Return(false);
     public bool AutoAdvance => false;
