+++ samples/TestApp/TestApp/Samples/ControlsNew/Wizard/WizardViewModel.cs
@@ -5,14 +5,13 @@ using CSharpFunctionalExtensions;
 using ReactiveUI;
 using TestApp.Samples.ControlsNew.Wizard.Pages;
 using Zafiro.Avalonia.Dialogs;
+using Zafiro.Avalonia.Dialogs.Wizards.Classic;
 using Zafiro.CSharpFunctionalExtensions;
+using Zafiro.UI.Wizards.Classic.Builder;
 using FirstPageViewModel = TestApp.Samples.ControlsNew.Wizard.Pages.FirstPageViewModel;
 using SecondPageViewModel = TestApp.Samples.ControlsNew.Wizard.Pages.SecondPageViewModel;
 
 namespace TestApp.Samples.ControlsNew.Wizard;
-
-using FirstPageViewModel = FirstPageViewModel;
-using SecondPageViewModel = SecondPageViewModel;
 
 public class WizardViewModel : ReactiveObject
 {
@@ -24,13 +23,13 @@ public class WizardViewModel : ReactiveObject
             int number = 0;
             string? text = "";
 
-            var wizard = Zafiro.Avalonia.Controls.Wizards.Builder.WizardBuilder
+            var wizard = WizardBuilder
                 .StartWith(() => new FirstPageViewModel())
                 .Then(prev => new SecondPageViewModel(prev.Number!.Value), x => number = x.Number!.Value)
                 .Then(prev => new ThirdPageViewModel(prev), x => text = x.Text)
                 .FinishWith(prev => (prev.SomeProperty, number, text));
 
-            return dialog.ShowWizard(wizard, "Such a nice wizard this is!");
+            return dialog.ShowWizard<(string SomeProperty, int number, string text)>(wizard, "Such a nice wizard this is!");
         });
 
         LaunchWizard.Values()
