+++ src/Zafiro.Avalonia.Dialogs/Styles.axaml
@@ -1,10 +1,10 @@
 <Styles
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns="https://github.com/avaloniaui"
-        xmlns:dialogs="clr-namespace:Zafiro.Avalonia.Dialogs">
+    xmlns="https://github.com/avaloniaui"
+    xmlns:views="clr-namespace:Zafiro.Avalonia.Dialogs.Views">
 
     <Design.PreviewWith>
-        <dialogs:DialogView Width="500" Height="500" />
+        <views:DialogView Width="500" Height="500" />
     </Design.PreviewWith>
 
     <StyleInclude Source="avares://Zafiro.Avalonia/Styles.axaml" />
