--- src/Zafiro.Avalonia/Controls/SlimWizard/SlimWizardControl.axaml.cs
@@ -1,16 +1,0 @@
-using Avalonia.Controls.Primitives;
-using Zafiro.UI.Wizards;
-
-namespace Zafiro.Avalonia.Controls.SlimWizard;
-
-public class SlimWizardContorl : TemplatedControl
-{
-    public static readonly StyledProperty<ISlimWizard> WizardProperty = AvaloniaProperty.Register<SlimWizardContorl, ISlimWizard>(
-        nameof(Wizard));
-
-    public ISlimWizard Wizard
-    {
-        get => GetValue(WizardProperty);
-        set => SetValue(WizardProperty, value);
-    }
-}
