--- src/Zafiro.Avalonia/Controls/Wizards/Builder/IBusy.cs
@@ -1,6 +1,0 @@
-namespace Zafiro.Avalonia.Controls.Wizards.Builder;
-
-public interface IBusy
-{
-    public IObservable<bool> IsBusy { get; }
-}
