--- src/Zafiro.Avalonia/Controls/Wizards/Builder/IStep.cs
@@ -1,9 +1,0 @@
-using CSharpFunctionalExtensions;
-
-namespace Zafiro.Avalonia.Controls.Wizards.Builder;
-
-public interface IStep : IValidatable, IBusy
-{
-    bool AutoAdvance { get; }
-    Maybe<string> Title => Maybe<string>.None;
-}
