+++ src/Zafiro.Avalonia/Controls/Wizards/Classic/WizardHeader.cs
@@ -2,12 +2,21 @@ using Avalonia.Controls.Primitives;
 using CSharpFunctionalExtensions;
 using Zafiro.UI.Commands;
 
-namespace Zafiro.Avalonia.Controls.Wizards;
+namespace Zafiro.Avalonia.Controls.Wizards.Classic;
 
 public class WizardHeader : TemplatedControl
 {
     public static readonly StyledProperty<Maybe<string>> TitleProperty = AvaloniaProperty.Register<WizardHeader, Maybe<string>>(
         nameof(Title));
+
+    public static readonly StyledProperty<int> CurrentPageIndexProperty = AvaloniaProperty.Register<WizardHeader, int>(
+        nameof(CurrentPageIndex));
+
+    public static readonly StyledProperty<int> TotalPagesProperty = AvaloniaProperty.Register<WizardHeader, int>(
+        nameof(TotalPages));
+
+    public static readonly StyledProperty<IEnhancedCommand> BackProperty = AvaloniaProperty.Register<WizardHeader, IEnhancedCommand>(
+        nameof(Back));
 
     public Maybe<string> Title
     {
@@ -15,26 +24,17 @@ public class WizardHeader : TemplatedControl
         set => SetValue(TitleProperty, value);
     }
 
-    public static readonly StyledProperty<int> CurrentPageIndexProperty = AvaloniaProperty.Register<WizardHeader, int>(
-        nameof(CurrentPageIndex));
-
     public int CurrentPageIndex
     {
         get => GetValue(CurrentPageIndexProperty);
         set => SetValue(CurrentPageIndexProperty, value);
     }
 
-    public static readonly StyledProperty<int> TotalPagesProperty = AvaloniaProperty.Register<WizardHeader, int>(
-        nameof(TotalPages));
-
     public int TotalPages
     {
         get => GetValue(TotalPagesProperty);
         set => SetValue(TotalPagesProperty, value);
     }
-
-    public static readonly StyledProperty<IEnhancedCommand> BackProperty = AvaloniaProperty.Register<WizardHeader, IEnhancedCommand>(
-        nameof(Back));
 
     public IEnhancedCommand Back
     {
