+++ src/Zafiro.Avalonia/Controls/Wizards/Slim/SlimWizardControl.axaml.cs
@@ -1,0 +1,16 @@
+using Avalonia.Controls.Primitives;
+using Zafiro.UI.Wizards.Slim;
+
+namespace Zafiro.Avalonia.Controls.Wizards.Slim;
+
+public class SlimWizardControl : TemplatedControl
+{
+    public static readonly StyledProperty<ISlimWizard> WizardProperty = AvaloniaProperty.Register<SlimWizardControl, ISlimWizard>(
+        nameof(Wizard));
+
+    public ISlimWizard Wizard
+    {
+        get => GetValue(WizardProperty);
+        set => SetValue(WizardProperty, value);
+    }
+}
