+++ src/Zafiro.Avalonia/Controls/Wizards/Slim/WizardDesign.cs
@@ -1,7 +1,7 @@
 using Zafiro.UI.Commands;
-using Zafiro.UI.Wizards;
+using Zafiro.UI.Wizards.Slim;
 
-namespace Zafiro.Avalonia.Controls.SlimWizard;
+namespace Zafiro.Avalonia.Controls.Wizards.Slim;
 
 public class WizardDesign : ISlimWizard
 {
@@ -9,11 +9,4 @@ public class WizardDesign : ISlimWizard
     public IEnhancedCommand Back { get; } = ReactiveCommand.Create(() => { }).Enhance();
     public IPage CurrentPage { get; } = new PageDesign();
     public int TotalPages { get; } = 3;
-}
-
-public class PageDesign : IPage
-{
-    public object Content { get; } = "This is some content";
-    public string Title { get; } = "Title";
-    public int Index { get; } = 2;
 }
