--- src/Zafiro.Avalonia/Controls/Wizards/WizardTemplate.axaml
@@ -1,10 +1,0 @@
-<DataTemplates xmlns="https://github.com/avaloniaui"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:w="clr-namespace:Zafiro.Avalonia.Controls.Wizards"
-                    x:CompileBindings="True">
-    
-    <DataTemplate DataType="w:IWizard">
-        <w:WizardControl Wizard="{Binding}" />
-    </DataTemplate>
-
-</DataTemplates>
